requirements.txt
django==1.11.3
mysqlclient
djangorestframework